Restructure package, datascience.py -> table.py, and stub map widget
Before, all the code in this package lived in the `datascience.py` file.

In this commit, I add a folder structure to the package to facilitate
adding additional functionality. Specifically, I add a `maps` module
which will eventually contain the logic to render a GMap widget in a
notebook (see #5).

The widget will require Javascript to be placed in the user's
`nbextension` directory, so I add the relevant code in `setup.py` to do
so.

After this commit is merged, we are still to type:

```python

from datascience import *

letters = ['a', 'b', 'c', 'z']
counts = [9, 3, 3, 1]
points = [1, 2, 2, 10]
t = Table([('letter', letters), ('count', counts), ('points', points)])
print(t)
letter | count | points
a      | 9     | 1
b      | 3     | 2
c      | 3     | 2
z      | 1     | 10

```

With an eventual API of

```python

from datascience import draw_map

draw_map(...)
```

Resolves #15