Add folium integration and use it in an example notebook
With the discovery of the folium package, I remove the old hacky stuff
I put in before for gmaps and use folium instead!

I also added an example notebook in the `examples/` folder that
successfully renders a map centered at Broad Street (now Broadwick
Street).

The next step is to add the points / polygon functionality.