Provide new minimal representation for row
A row was earlier represented as a namedtuple. But this resulted in
the restriction that a table could not have more than 255 columns
(see http://stackoverflow.com/questions/18550270/any-way-to-bypass-namedtuple-255-arguments-limitation).

The new representation is a subclass of tuple which still
approximately provides access to column labels and has the same
representation. Additionally, column values can be accessed as
attributes and a mapping from column label to values can be obtained
using the asdict() method. Other behavior can be filled in as
required.