Install packages through conda instead of pip
This commit resolves the pain we had with needing scipy in the
datascience package. Travis has a lot of problems installing it,
apparently.

Following the suggestions at
http://lmjohns3.com/2015/06/using-travis-ci-with-miniconda-scipy-and-nose.html,
I install Miniconda and then install the necessary packages through
conda instead of pip.

This allows the packages to be easily installed and it looks like Travis
can cache the installed packages for even faster build times.